make file util for custom table component
